userbot/requirements.txt
python-dotenv==1.1.1
Telethon==1.40.0
